color: Clamp extended channels at one-dimensional curves
Plane color pipelines preserve signed, extended-range channel values
between matrix operations and apply one-dimensional curves through
normalized lookup tables.

Passing an extended matrix result to an unsigned curve interface wraps
negative values and values above one. The following curve then samples an
unrelated part of its lookup table instead of the nearest endpoint.

This commit resolves the domain mismatch by accepting signed curve inputs
and clamping each channel before lookup and interpolation. A focused KUnit
case covers negative and over-range matrix output entering an sRGB curve.

The color pipeline now retains extended precision between matrix operations
while giving one-dimensional curves defined endpoint behavior.

+++ src/castkms_composer.h
@@ -27,8 +27,9 @@ bool castkms_composer_demand_put(struct castkms_composer_demand *demand,
 				 enum castkms_composer_client client);
 u16 castkms_lerp_u16(u16 a, u16 b, s64 t);
 s64 castkms_get_lut_index(const struct castkms_color_lut *lut, u16 channel_value);
-u16 castkms_apply_lut_to_channel_value(const struct castkms_color_lut *lut, u16 channel_value,
-			       enum lut_channel channel);
+u16 castkms_apply_lut_to_channel_value(const struct castkms_color_lut *lut,
+				       s32 channel_value,
+				       enum lut_channel channel);
 void castkms_apply_3x4_matrix(struct pixel_argb_s32 *pixel, const struct drm_color_ctm_3x4 *matrix);
 #endif
